Refactor/Contract helpers
