+++ src/modules/contracts/hooks/useContractInstance.ts
@@ -2,42 +2,38 @@ import { useWeb3 } from 'modules/blockChain/hooks/useWeb3'
 import { useCurrentChain } from 'modules/blockChain/hooks/useCurrentChain'
 import { useGlobalMemo } from 'shared/hooks/useGlobalMemo'
 import {
-  ConnectRpcFn,
-  ConnectWeb3Fn,
+  ContractConnector,
   ContractFactoryAbstract,
   FactoryInstance,
 } from '../types'
-import type { ChainAddress } from '../utils/BunchAddresses'
 
 export function useContractInstanceRpc<F extends ContractFactoryAbstract>(
-  chainAddres: ChainAddress,
-  connectRpc: ConnectRpcFn<F>,
+  connector: ContractConnector<F>,
 ) {
   type Instance = FactoryInstance<F>
 
   const chainId = useCurrentChain()
-  const address = chainAddres.get(chainId)
+  const address = connector.chainAddress.get(chainId)
 
   return useGlobalMemo(
-    () => connectRpc({ chainId }) as Instance,
+    () => connector.connectRpc({ chainId }) as Instance,
     `contract-rpc-${address}`,
   )
 }
 
 export function useContractInstanceWeb3<F extends ContractFactoryAbstract>(
-  chainAddres: ChainAddress,
-  connectWeb3: ConnectWeb3Fn<F>,
+  connector: ContractConnector<F>,
 ) {
   type Instance = FactoryInstance<F>
 
   const { library, active, account } = useWeb3()
   const chainId = useCurrentChain()
   const activeKey = active ? 'active' : 'inactive'
-  const address = chainAddres.get(chainId)
+  const address = connector.chainAddress.get(chainId)
 
   return useGlobalMemo(
     () =>
-      connectWeb3({
+      connector.connectWeb3({
         chainId,
         library: library?.getSigner(),
       }) as Instance,
